refactor: move container to the index level
