+++ __tests__/e2e/root.e2e.spec.ts
@@ -2,9 +2,9 @@ import '../mocks/config.provider.mock.js';
 import { ServerType } from '@hono/node-server';
 import { Db, DbProvider } from 'src/secondary/adapters/db.provider.js';
 import { Browser, chromium, Page } from 'playwright';
-import { Server } from 'src/primary/server.js';
+import { Server } from 'src/primary/adapters/server.js';
 import { http, HttpResponse, JsonBodyType } from 'msw';
-import { Container } from 'src/primary/adapters/container.js';
+import { Container } from 'src/container.js';
 import { createMockServer } from '__tests__/utils/mock-server.utils.js';
 import { CacheProvider } from 'src/secondary/adapters/cache.provider.js';
 import { Chance } from 'chance';
