+++ __tests__/integration/metrics.controller.spec.ts
@@ -1,6 +1,6 @@
 import { ctx } from '__tests__/setup-integration-context.js';
 import { HttpStatus } from 'src/secondary/types/http.types.js';
-import { Server } from 'src/primary/server.js';
+import { Server } from 'src/primary/adapters/server.js';
 import { MetricsProvider } from 'src/secondary/adapters/metrics.provider.js';
 
 describe('metrics controller (integration)', () => {
