+++ __tests__/setup-integration-context.ts
@@ -1,5 +1,5 @@
 import './mocks/index.js';
-import { Container } from 'src/primary/adapters/container.js';
+import { Container } from 'src/container.js';
 
 export const ctx = new Container();
